+++ openedx_events/tests/test_producer_config.py
@@ -31,7 +31,7 @@ class ProducerConfiguratonTest(TestCase):
     @patch('openedx_events.apps.get_producer')
     def test_enabled_disabled_events(self, mock_producer):
         """
-        Check whether XBLOCK_PUBLISHED is connected to the handler and the handler only publishes enabled events.
+        Check whether XBLOCK_PUBLISHED is connected to the handler and the handler only produces enabled events.
 
         Args:
             mock_producer: mock get_producer to inspect the arguments.
@@ -46,12 +46,12 @@ class ProducerConfiguratonTest(TestCase):
         # check that call_args_list only consists of enabled topics.
         call_args = mock_send.send.call_args_list[0][1]
         self.assertDictContainsSubset(
-            {'topic': 'content-authoring-xblock-lifecycle', 'event_key_field': 'xblock_info.usage_key'},
+            {'topic': 'enabled_topic_a', 'event_key_field': 'xblock_info.usage_key'},
             call_args
         )
         call_args = mock_send.send.call_args_list[1][1]
         self.assertDictContainsSubset(
-            {'topic': 'content-authoring-all-status', 'event_key_field': 'xblock_info.usage_key'},
+            {'topic': 'enabled_topic_b', 'event_key_field': 'xblock_info.usage_key'},
             call_args
         )
 
@@ -78,21 +78,23 @@ class ProducerConfiguratonTest(TestCase):
             with pytest.raises(ProducerConfigurationError, match="should be a dictionary"):
                 apps.get_app_config("openedx_events").ready()
 
-        with override_settings(EVENT_BUS_PRODUCER_CONFIG={"invalid.event.type": []}):
+        with override_settings(EVENT_BUS_PRODUCER_CONFIG={"invalid.event.type": {}}):
             with pytest.raises(ProducerConfigurationError, match="No OpenEdxPublicSignal of type"):
                 apps.get_app_config("openedx_events").ready()
 
         with override_settings(EVENT_BUS_PRODUCER_CONFIG={"org.openedx.content_authoring.xblock.deleted.v1": ""}):
-            with pytest.raises(ProducerConfigurationError, match="should be a list or a tuple"):
+            with pytest.raises(ProducerConfigurationError, match="should be a dict"):
                 apps.get_app_config("openedx_events").ready()
 
-        with override_settings(EVENT_BUS_PRODUCER_CONFIG={"org.openedx.content_authoring.xblock.deleted.v1": [""]}):
-            with pytest.raises(ProducerConfigurationError, match="object is not a dictionary"):
+        with override_settings(EVENT_BUS_PRODUCER_CONFIG={"org.openedx.content_authoring.xblock.deleted.v1":
+                                                          {"topic": ""}}):
+            with pytest.raises(ProducerConfigurationError, match="One of the configuration objects is not a"
+                                                                 " dictionary"):
                 apps.get_app_config("openedx_events").ready()
 
         with override_settings(
             EVENT_BUS_PRODUCER_CONFIG={
-                "org.openedx.content_authoring.xblock.deleted.v1": [{"topic": "some", "enabled": True}]
+                "org.openedx.content_authoring.xblock.deleted.v1": {"topic": {"enabled": True}}
             }
         ):
             with pytest.raises(ProducerConfigurationError, match="missing 'event_key_field' key."):
@@ -100,9 +102,10 @@ class ProducerConfiguratonTest(TestCase):
 
         with override_settings(
             EVENT_BUS_PRODUCER_CONFIG={
-                "org.openedx.content_authoring.xblock.deleted.v1": [
-                    {"topic": "some", "enabled": 1, "event_key_field": "some"}
-                ]
+                "org.openedx.content_authoring.xblock.deleted.v1":
+                {
+                    "some": {"enabled": 1, "event_key_field": "some"}
+                }
             }
         ):
             with pytest.raises(